[add] 食事招待の情報詳細ページを作成し、デザインも完了。
