+++ frontend/assets/styles/common.css
@@ -68,3 +68,9 @@ header .el-menu-item {
     font-size: 50px;
   }
 } */
+
+.flex-center {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
